+++ app/src/main/java/com/le/potato/transport/HIDTransport.kt
@@ -5,7 +5,7 @@ import android.content.Context
 
 interface HIDTransport {
     fun init(context: Context, reportMap: ByteArray)
-    fun addInputReport(reportId: Int, report: ByteArray)
+    fun addInputReport(reportId: Byte, report: ByteArray)
     fun disconnect(device: BluetoothDevice)
     fun deactivate()
     val devices: Set<BluetoothDevice>
